legacy_xml asr_settings_from_xml_nodes
