+++ src/auto_splitter_settings.rs
@@ -4,9 +4,7 @@ use asr::future::retry;
 use std::path::Path;
 use xmltree::{Element, XMLNode};
 
-use ugly_widget::radio_button::options_str;
-
-use crate::{file, legacy_xml, splits::Split};
+use crate::{file, legacy_xml};
 
 pub async fn wait_asr_settings_init() -> asr::settings::Map {
     let settings1 = asr::settings::Map::load();
@@ -42,19 +40,8 @@ fn asr_settings_from_xml_string(xml_string: &str) -> Option<asr::settings::Map> 
 }
 
 fn asr_settings_from_xml_nodes(xml_nodes: Vec<XMLNode>) -> Option<asr::settings::Map> {
-    let splits = legacy_xml::splits_from_xml_nodes(xml_nodes)?;
-    Some(asr_settings_from_splits(&splits))
-}
-
-fn asr_settings_from_splits(splits: &[Split]) -> asr::settings::Map {
-    // new empty map, which will only include the new splits
-    let settings_map = asr::settings::Map::new();
-    let l = asr::settings::List::new();
-    for split in splits.iter() {
-        l.push(options_str(split));
-    }
-    settings_map.insert("splits", l);
-    settings_map
+    // TODO: deal with either Legacy XML or ASR XML
+    legacy_xml::asr_settings_from_xml_nodes(xml_nodes)
 }
 
 fn file_find_auto_splitter_settings<P: AsRef<Path>>(path: P) -> Option<Vec<XMLNode>> {
