+++ src/legacy_xml.rs
@@ -1,7 +1,24 @@
 
+use ugly_widget::radio_button::options_str;
 use xmltree::XMLNode;
 
 use crate::splits::Split;
+
+pub fn asr_settings_from_xml_nodes(xml_nodes: Vec<XMLNode>) -> Option<asr::settings::Map> {
+    let splits = splits_from_xml_nodes(xml_nodes)?;
+    Some(asr_settings_from_splits(&splits))
+}
+
+fn asr_settings_from_splits(splits: &[Split]) -> asr::settings::Map {
+    // new empty map, which will only include the new splits
+    let settings_map = asr::settings::Map::new();
+    let l = asr::settings::List::new();
+    for split in splits.iter() {
+        l.push(options_str(split));
+    }
+    settings_map.insert("splits", l);
+    settings_map
+}
 
 #[derive(Clone, Debug)]
 struct XMLSettings {
